[SPARK-8563] [MLlib] Fixed a build error.

+++ mllib/src/main/scala/org/apache/spark/mllib/linalg/distributed/IndexedRowMatrix.scala
@@ -146,7 +146,7 @@ class IndexedRowMatrix(
       val indexedRows = indices.zip(svd.U.rows).map { case (i, v) =>
         IndexedRow(i, v)
       }
-      new IndexedRowMatrix(indexedRows, svd.U.numRows(), svd.U.numCols())
+      new IndexedRowMatrix(indexedRows, svd.U.numRows(), svd.U.numCols().toInt)
     } else {
       null
     }
